Add new configuration helper script
The script will take on the functionality which was previously available
via the `.f` command. In the future, an alias will be added to that
command so that the script can be invoked via `.f config`.

Future commits will involve removing duplicate functionality inside the
`.f` command, as well as integrating the configuration tool into the
installation process.